+++ src/Field_Base.php
@@ -79,13 +79,6 @@ class Field_Base {
 	private Setting|false $setting = false;
 
 	/**
-	 * The dataview type.
-	 *
-	 * @var string
-	 */
-	protected string $dataview_type = 'text';
-
-	/**
 	 * Constructor.
 	 *
 	 * @param Settings $settings_obj The settings object.